Add importable wads.repo_audit (SSOT classifier); skill script -> shim
Promotes the repo-audit logic from the wads-repo-doctor skill's bundled script
to an importable module, wads/repo_audit.py, with a clean
audit_repo(repo, *, network=True) -> dict entry point (plus `python -m
wads.repo_audit`). The skill's scripts/repo_audit.py becomes a thin shim that
forwards to the module, so the documented invocation keeps working; SKILL.md
points at the module form.

Makes the "what is a modern wads repo" classification a single source of truth
shared by the skill, priv.fleet_status, and CI — instead of a copy living under
a skills directory that consumers had to shell out to via a machine-specific path.
